Force tiles to be a multiple of 2**depth when using UNet.
Translational invariance is broken in UNets, due to the pooling layers. To avoid artefacts in during prediction, `tile_size` are now forced to be a multiple of 2**depth.

This does not impact the "Lightning API", and can be an issue for small Z depth during prediction. In such a case, padding can be added or using prediction while by passing the CAREamist.