Make normalization mandatory (#139)
### Description

The current code base applies denormalization automatically (in the
prediction step of the `LightningModule` and in the BMZ export), and the
datasets all compute the statistics of the data (mean and std). But
`Normalization` can be disabled through the Pydantic model!

This PR makes `Normalization` mandatory by adding it to the list of
transforms directly in the datasets. This simplifies the configuration
(no more adding mean/std in two places) and ensures that normalization
is always applied.

One way to still disable it, is to choose 0 mean and 1 std.


- **What**: Make normalization mandatory in the code.
- **Why**: Normalization is understood in the code as being always
applied, yet it can be disabled through the configuration.
- **How**: `Normalization` is removed from the list of transforms and
automatically added when creating the `Compose` in the datasets.

### Changes Made

- **Modified**: `DataModel`, `SupportedTransforms`, the datasets.
- **Removed**: `NormalizationModel` from `DataModel` and
`SupportedTransforms`.

### Breaking changes

Any code specifically calling `Normalization` in the configuration.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)