Feature: Allow Noise2Noise with different in and out channels (#152)
### Description

Until now, Noise2Noise was only possible with the same number of input
and output channels. But one can envision have multiple channels in
(e.g. time frames) and supervising with a single frame.

This PR relaxes the constraints on Noise2Noise: channels in and out can
now be different, and UNet is not mandatory!

- **What**: Remove `algorithm_model` constraint on `N2N`, and change the
configuration convenience function.
- **Why**: We might try to run Noise2Noise with different in and out
channels.
- **How**: See what.

### Changes Made

- **Modified**: `algorithm_model` and `create_n2n_configuration`, and
the corresponding tests.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [x] PR to the documentation exists (for bug fixes / features)
    - https://github.com/CAREamics/careamics.github.io/pull/9
    - https://github.com/CAREamics/careamics-examples/pull/3