Feature: Use current weights for prediction (#166)
### Description

Currently, we pass both `model` and `checkpoint` to `trainer.predict`.
My understanding of the
[docs](https://lightning.ai/docs/pytorch/stable/common/trainer.html#predict)
is that `checkpoint` takes precedence.

It is not clear what model is the best to use (`last` vs `best`),
especially in the context of Noise2Void. In order to avoid a behaviour
that has not be clarified and described, this PR removes the
`checkpoint` parameter altogether.

In the future, we might want to revisit that. Currently, we would
advocate letting users configure the checkpoint callback and then load
the correct weights themselves.

- **What**: Remove `checkpoint` parameter from `careamist.predict`.
- **Why**: Clarify which weights are used (in this case, the current
ones).
- **How**: Remove the parameter from the `predict` function signature,
and from the call to the `trainer`.

### Changes Made


- **Modified**: `careamist.py`.


---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)