Feature: Prediction writer callback initialisation (#179)
### Description

Add the skeleton and initialisation of the `PredictionWriterCallback`
class, to allow users to save predictions to disk in the future. This
callback will be used by the `CAREamist` class but should also be
compatible with users preferring to use the internal `lightning` API.

- **What**: 
  - Add `PredictionWriterCallback` class.
  - Implement `PredictionWriterCallback.__init__` method.
- Add placeholder for `write_on_batch_end` hook, with docstring
description of what is coming.
- Implement `setup` hook that will create the prediction output
directory when the `lightning` predict stage starts.
  - Tests for `PredictionWriterCallback.__init__`.
- **Why**: 
- This callback will be used to save predictions to disk, this is a
feature in the works for a while. This will also allow the predict
command to be implemented in the CLI.

### Changes Made

- **Added**: 
- `PredictionWriterCallback` class to the `lightning.callbacks` package.
  - Tests for the initialisation of `PredictionWriterCallback`.
- **Modified**: 
  - Package/sub-package level imports (__init__.py files.)

### Additional Notes and Examples

#### Future considerations

- There will have to be a method to update `save_type`,
`save_extension`, `save_func` & `save_func_kwargs` because in the
`CAREamist` class they will be decided after the initialisation of
`PredictionWriterCallback`. The `CAREamist` class will have a
`predict_to_disk` method that where users can pass these parameters, but
the `PredictionWriterCallback` will have to be initialised during
`CAREamist` initialisation because it needs to be passed to the trainer.
- For tiling/not tiling & tile caching/writing tiles directly to disk
(`zarr`). I want to follow a [strategy
pattern](https://refactoring.guru/design-patterns/strategy) so
`PredictionWriterCallback` will interact with a strategy interface and
it's only responsibility will be to select the correct one.
- As mentioned in the description, this callback should be easily usable
by both the `CAREamist` class and users who are using the `lightning`
API.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)