fix: fixing various bugs to enable microSplit training  (#242)
### Description

This PR aims at solving a variety of bugs present throughout the LVAE
codebase to allow training and evaluation of microSplit family of
models. Examples of how to train the different versions of microSplit
are available [in this
repo](https://github.com/CAREamics/microSplit-reproducibility).
Also added comments about things to be dealt with in future iterations.

- **What**: Fixed bugs in different places interfering with microSplit
training and evaluation. Now it is possible to train the model.
- **Why**: Of course we need a way to train and evaluate the models ;)
- **How**: Depending on the particular bug encountered.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [ ] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: CatEek <zubarev.ia@gmail.com>
Co-authored-by: melisande-c <milly.croft@gmail.com>