Add stop training method (#249)
### Description

If CAREamics is used in a multithreaded application (e.g. napari
plugin), then we need a way to stop the training programmatically. This
PR add a `stop_training` method to the `CAREamist`.

Additionally, this PR removes the unused parameter `experiment_name`
from the CAREamist.

- **What**: Add stop training method.
- **Why**: Enable stopping training in multithreaded applications (e.g.
napari)
- **How**: Add `stop_training` method that raises the `should_stop` flag
of Lightning.

### Changes Made

- **Modified**: `careamist.py`, by adding the `stop_training` function
and cleaning up unused keywords.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)