Feature: Predict to disk (outerloop implementation) (#253)
### Description

- **What**: Add a `predict_to_disk` function to the `CAREamist` class.
- **Why**: So users can save predictions without having to write the
saving process themselves.
- **How**: This implementation loops through the files, predicts the
result and saves each one in turn. N.b. this will eventually be replaced
with the `PredictionWriterCallback` version.

### Changes Made

- **Added**: 
  - `predict_to_disk` function to `CAREamist` class.
  - tests
- **Modified**:
  - Some type hints.

### Additional Notes and Examples

Currently the results are saved to a directory called "predictions" but
maybe it should be configurable by the user?

This function can only be called on source data from a path because the
prediction files are saved with the same name as the source files.

Not the neatest as I expect this to be replaced soon.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)