refac: added possibility to pick `kl_restricted` loss & other loss-related refactoring (#272)
### Description

In some LVAE training examples (see `microSplit_reproducibility` repo)
there is the need to consider the *restricted KL loss*, instead of the
simple sample-wise one. This PR allows the user to pick that one.

- **What**: The KL loss type is no longer hardcoded in the loss
functions. Now it is possible to pick also the `restricted_kl` KL loss
type.
- **Why**: It is needed for some experiments/examples.
- **How**:  added an input parameter to the KL loss functions.

### Breaking changes

The `kl_type` parameter is added in the loss functions, so we need to be
careful of correctly specifying it in the examples in the
`microSplit_reproducibility` repo.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)