Feature: Add predict command to CLI (#281)
### Description

- **What**: Implement CLI predict command.
- **Why**: Previously calling the CLI predict command would raise
`NotImplementedError`.
- **How**: A given checkpoint is loaded, then the
`CAREamist.predict_to_disk` is called with optional parameters. Using
Typer.

### Changes Made

- **Added**:
  - CLI predict command in `careamics/cli/main.py`.
  - Tests for CLI predict command.
- **Modified**:
- `CAREamist.predict_to_disk`: fixed bug relating to prediction output
directory structure for a single file + some reorganisation.
- For CLI replaced `patch_size_callback` with `handle_2D_3D_callback`,
moved to `careamics/cli/utils.py`. (Typer can only handle fixed sized
inputs so this callback is used for a parameter like `tile_size` that
can be 2D or 3D).
- `predict_to_disk` tests. Predictions are saved directly to the
selected predictions output directory (Not in a second level internal
directory).

### Additional Notes and Examples

Calling the train and predict command (for a N2V config) is as follows

```
$ careamics train <CONFIG PATH> -ts <TRAIN DATA PATH>
$ careamics predict checkpoints/<EXPERIMENT NAME>.ckpt <TEST DATA PATH>
```

Also, please note that `CAREamist.predict_to_disk` is quite messy but it
should be replaced soon when the `PredictionWriterCallback is complete.

---

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)

---------

Co-authored-by: melisande-c <melisande.croft@fht.org>